a3-orchestrator/runtime/requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
numpy==1.26.3
outcome==1.3.0.post0
pillow==10.2.0
PySocks==1.7.1
requests==2.31.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
